Refactor animations so now the cards cover the whole screen, and refactor code to make it cleaner
