+++ godtools/Views/TractElements/HeaderAnimations.swift
@@ -1,0 +1,32 @@
+//
+//  HeaderAnimations.swift
+//  godtools
+//
+//  Created by Devserker on 5/4/17.
+//  Copyright © 2017 Cru. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+extension Header {
+    
+    func showHeader() {
+        UIView.animate(withDuration: 0.35,
+                       delay: 0.0,
+                       options: UIViewAnimationOptions.curveEaseInOut,
+                       animations: {
+                        self.transform = CGAffineTransform(translationX: 0, y: 0.0) },
+                       completion: nil )
+    }
+    
+    func hideHeader() {
+        UIView.animate(withDuration: 0.35,
+                       delay: 0.0,
+                       options: UIViewAnimationOptions.curveEaseInOut,
+                       animations: {
+                        self.transform = CGAffineTransform(translationX: 0, y: -self.yPosition) },
+                       completion: nil )
+    }
+    
+}
