+++ godtools/Views/TractElements/Heading.swift
@@ -25,19 +25,21 @@ class Heading: BaseTractElement {
         self.backgroundColor = .green
     }
     
-    fileprivate func buildFrame() -> CGRect {
-        return CGRect(x: self.xPosition,
-                      y: self.yPosition,
-                      width: self.width,
-                      height: self.height)
-    }
-    
     override func textStyle() -> (style: String, width: CGFloat, height: CGFloat) {
         return ("toolFrontTitle", BaseTractElement.Standards.textContentWidth, CGFloat(0.0))
     }
     
     override func yEndPosition() -> CGFloat {
         return self.yPosition + self.height
+    }
+    
+    // MARK: - Helpers
+    
+    fileprivate func buildFrame() -> CGRect {
+        return CGRect(x: self.xPosition,
+                      y: self.yPosition,
+                      width: self.width,
+                      height: self.height)
     }
 
 }
