+++ godtools/Views/TractElements/TractLabel.swift
@@ -51,35 +51,13 @@ class TractLabel: BaseTractElement {
         return self.yStartPosition + self.height
     }
     
+    // MARK: - Helpers
+    
     fileprivate func buildFrame() -> CGRect {
         return CGRect(x: self.xPosition,
                       y: self.yPosition,
                       width: self.width,
                       height: self.height)
-    }
-    
-    func setupPressGestures() {
-        if (self.parent?.isKind(of: Card.self))! {
-            let tapGesture = UITapGestureRecognizer(target: self, action: #selector(handleGesture))
-            tapGesture.numberOfTapsRequired = 1
-            tapGesture.numberOfTouchesRequired = 1
-            
-            let frame = CGRect(x: 0.0,
-                               y: 0.0,
-                               width: self.width,
-                               height: 60.0)
-            self.tapView.frame = frame
-            self.tapView.addGestureRecognizer(tapGesture)
-            self.tapView.isUserInteractionEnabled = true
-            self.addSubview(self.tapView)
-        }
-    }
-    
-    func handleGesture(sender: UITapGestureRecognizer) {
-        if sender.state == .ended {
-            let cardView = self.parent as! Card
-            cardView.changeCardState()
-        }
     }
 
 }
